VendeeGlobal 2020/2021 - Streamming Data
The Challenge - Build a cloud-based Lambda Architecture to process the telemetry data from the sailing boats. Your architecture should run in Azure and contain a real-time path for processing sailing boat data in real-time, and a batch-processing path for collecting sailing boat data in batches and performing calculations on those batches.